Improve the way to get the result and notification
Check the way how to catch the result from a the call Update.
It returns the couple (result, err) in a synchronius fashion
regarding the fact that the internal call is being executed inside
a coroutine.

Notice that this solution offers two ways to extract the result.
- By notification
- By returning value

The motivation for that is that a Query() should return a result
for whom it may concern. E.g. The result of a Query() call.
In the other hand, an action like Update() leads to a result that
should be notified the users interested on that modification.